# openapi.yaml
openapi: 3.0.3
info:
  title: Nuts Knooppunt API
  description: >
    API served on the Knooppunt's internal interface (default `:8081`).

    See [docs/INTEGRATION.md](https://github.com/nuts-foundation/nuts-knooppunt/blob/main/docs/INTEGRATION.md) for narrative documentation.

    FHIR resource bodies are typed by resource (Bundle, List, Subscription,
    OperationOutcome) but not redefined in full — each schema links to its
    definition in the FHIR R4 spec.

    Out of scope here: `/mitz/notify`, the public-interface endpoint MITZ
    calls directly (not integrated against by Knooppunt users) — see
    [mitz-public.openapi.yaml](https://github.com/nuts-foundation/nuts-knooppunt/blob/main/mitz-public.openapi.yaml).
    Everything under `/nuts` is proxied straight through to the embedded
    Nuts node; see the
    [Nuts node API reference](https://nuts-node.readthedocs.io/en/project-gf/pages/integrating/api.html)
    for those endpoints.
  version: "1.0"
security: []
...
servers:
  - url: http://localhost:8081
    description: Internal interface (status, PDP, NVI, MITZ subscription)
...
paths:
  /status:
    get:
      tags: [status]
      summary: Readiness check
      operationId: getStatus
      description: >
        Returns `200 OK` once all components have started. Returns `503` with
        body `starting` while the Knooppunt is still starting up. Used as the
        de facto health check for this service (there is no separate `/health`
        endpoint on the Knooppunt itself).
      responses:
        "200":
          description: Ready
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        "503":
          description: Not yet ready
          content:
            text/plain:
              schema:
                type: string
                example: "starting"
...
  /lrza/update:
    post:
      tags: [lrza]
      summary: Trigger LRZA synchronization
      operationId: triggerLrzaSync
      description: >
        Synchronizes mCSD resources from the centralised LRZA directory into
        the local Query Directory. Runs synchronously and returns once the
        cycle completes; the Knooppunt does not sync on a schedule of its
        own, so this is meant to be triggered manually or via cron.
      responses:
        "200":
          description: Synchronization report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryUpdateReport"
        "500":
          description: Synchronization failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "Failed to update LRZA: connection refused"
  /pdp:
    post:
      tags: [pdp]
      summary: Evaluate the default authorization policy
      operationId: evaluateDefaultAuthorization
      description: >
        Equivalent to `POST /pdp/v1/data/knooppunt/authz` — same request
        body, responses and behavior; see that operation for details. This
        is a shorter path to the same, single policy that route's
        `knooppunt/authz` package/rule addresses via the Open Policy Agent
        convention.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PDPAuthzRequest"
      responses:
        "200":
          description: Authorization decision. See `POST /pdp/v1/data/knooppunt/authz` for details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDPAuthzResponse"
        "400":
          description: >
            Request body isn't valid JSON, or the scope names a disallowed
            policy (`test_*` or `system`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDPAuthzResponse"
  /pdp/v1/data/knooppunt/authz:
    post:
      tags: [pdp]
      summary: Evaluate an authorization request
      operationId: evaluateAuthorization
      description: >
        Evaluates whether a data-holder request should be allowed, based on
        subject/organization, the HTTP request that is being authorized
        (e.g. FHIR request), and its context (e.g. patient BSN, connection
        type).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PDPAuthzRequest"
            examples:
              mcsdUpdate:
                summary: mCSD update request
                value:
                  input:
                    subject:
                      organization_ura: "00000666"
                      scope: mcsd_update
                    request:
                      method: GET
                      protocol: "HTTP/1.1"
                      path: /Organization/_history
                    context:
                      data_holder_organization_id: "00000659"
                      connection_type_code: hl7-fhir-rest
              searchWithQueryParams:
                summary: Search request with a query parameter
                value:
                  input:
                    subject:
                      organization_ura: "00000666"
                      scope: mcsd_update
                    request:
                      method: GET
                      protocol: "HTTP/1.1"
                      path: /Organization
                      query: "_since=2024-01-01"
                    context:
                      data_holder_organization_id: "00000659"
                      connection_type_code: hl7-fhir-rest
              bgzPatientRead:
                summary: bgz scope reading a specific patient
                value:
                  input:
                    subject:
                      user_id: "000095254"
                      user_role: "01.015"
                      organization_ura: "00000666"
                      organization_facility_type: Z3
                      scope: bgz
                    request:
                      method: GET
                      protocol: "HTTP/1.1"
                      path: /Patient/3E439979-017F-40AA-594D-EBCF880FFD97
                    context:
                      data_holder_organization_id: "00000659"
                      data_holder_facility_type: Z3
                      connection_type_code: hl7-fhir-rest
              formEncodedSearch:
                summary: Search with form-encoded body (pzp_gf scope)
                value:
                  input:
                    subject:
                      user_id: "000095254"
                      user_role: "01.015"
                      organization_ura: "00000666"
                      organization_facility_type: Z3
                      scope: pzp_gf
                    request:
                      method: POST
                      protocol: "HTTP/1.1"
                      path: /Patient/_search
                      header:
                        Content-Type: ["application/x-www-form-urlencoded"]
                      body: "identifier=http://fhir.nl/fhir/NamingSystem/bsn|900214181"
                    context:
                      data_holder_organization_id: "90000697"
                      data_holder_facility_type: Z3
                      connection_type_code: hl7-fhir-rest
      responses:
        "200":
          description: >
            Authorization decision. Also returned, with `error` set and
            `allow` `false`, when the request body was valid JSON but
            semantically invalid (e.g. an unparseable FHIR request, or no
            scope given) or names a policy bundle that doesn't exist. Only
            an unparseable JSON body, or a scope naming a disallowed policy,
            gets a `400` instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDPAuthzResponse"
              examples:
                allowed:
                  value:
                    allow: true
                    policies:
                      bgz:
                        allow: true
                        reasons:
                          - code: info
                            description: MITZ consent granted
                denied:
                  value:
                    allow: false
                    policies:
                      bgz:
                        allow: false
                        reasons:
                          - code: not_allowed
                            description: No MITZ consent found for this patient
        "400":
          description: >
            Request body isn't valid JSON, or the scope names a disallowed
            policy (`test_*` or `system`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDPAuthzResponse"
              example:
                allow: false
                error: "unable to parse request body: unexpected end of JSON input"
                policies: {}
...
  /pdp/bundles/{policyName}:
    get:
      tags: [pdp]
      summary: Get a loaded policy bundle by name
      operationId: getAuthorizationPolicyBundle
      description: >
        Serves the raw Open Policy Agent (OPA) bundle (gzipped tar) for a
        given policy. Primarily used by the Knooppunt itself; may change
        or be removed without prior notice.
      parameters:
        - name: policyName
          in: path
          required: true
          description: Name of the loaded policy bundle, as returned by `GET /pdp/bundles`.
          schema:
            type: string
      responses:
        "200":
          description: Bundle contents
        "404":
          description: No bundle with this name is loaded
  /nvi:
    post:
      tags: [nvi]
      summary: Register a List via a transaction Bundle
      operationId: registerNVIListBundle
      description: >
        Submits a FHIR transaction Bundle containing one or more `List`
        resources. BSN values in `List.subject.identifier` are pseudonymized
        (tokenized) before the Bundle is forwarded to the upstream NVI FHIR
        server.
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: "#/components/schemas/FHIRBundle"
            example:
              resourceType: Bundle
              type: transaction
              entry:
                - resource:
                    resourceType: List
                    status: current
                    mode: working
                    subject:
                      identifier:
                        system: "http://fhir.nl/fhir/NamingSystem/bsn"
                        value: "999911136"
                    source:
                      identifier:
                        system: "http://fhir.nl/fhir/NamingSystem/ura"
                        value: "00000020"
                    code:
                      coding:
                        - system: "http://minvws.github.io/generiekefuncties-docs/CodeSystem/nl-gf-data-categories-cs"
                          code: MedicationRequest
                    extension:
                      - url: "http://minvws.github.io/generiekefuncties-docs/StructureDefinition/nl-gf-localization-custodian"
                        valueReference:
                          identifier:
                            system: "http://fhir.nl/fhir/NamingSystem/ura"
                            value: "00000020"
                  request:
                    method: POST
                    url: List
      responses:
        "200":
          description: Resulting transaction-response Bundle from NVI
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIRBundle"
              example:
                resourceType: Bundle
                type: transaction-response
                entry:
                  - resource:
                      resourceType: List
                      id: "1"
                      status: current
                      mode: working
                      subject:
                        identifier:
                          system: "http://fhir.nl/fhir/NamingSystem/bsn"
                          value: "999911136"
                    response:
                      status: "201 Created"
                      location: "List/1"
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
  /nvi/List:
    post:
      tags: [nvi]
      summary: Register a List directly
      operationId: registerNVIList
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: "#/components/schemas/FHIRList"
      responses:
        "200":
          description: Created List, as returned by NVI
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIRList"
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
    get:
      tags: [nvi]
      summary: Search for Lists
      operationId: searchNVILists
      description: >
        Requires at least one of `patient:identifier`, `subject:identifier` or
        `source:identifier`. `patient:identifier` is mapped to
        `subject:identifier` before forwarding, since NVI only supports the
        latter. BSN identifier values are tokenized before being sent upstream.
        An optional `code` parameter is also supported. If the upstream NVI
        response is paginated, this returns `422` since pagination isn't
        supported.
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: patient:identifier
          in: query
          required: false
          description: >
            BSN of the patient referenced by `List.subject`, as a FHIR
            token (`system|value`). Mapped to `subject:identifier` before
            forwarding to NVI.
          schema:
            type: string
          example: "http://fhir.nl/fhir/NamingSystem/bsn|999911136"
        - name: subject:identifier
          in: query
          required: false
          description: >
            BSN of the patient referenced by `List.subject`, as a FHIR
            token (`system|value`). NVI's native parameter name; equivalent
            to `patient:identifier`.
          schema:
            type: string
        - name: source:identifier
          in: query
          required: false
          description: >
            Identifier of `List.source` (the system that registered the
            List), as a FHIR token (`system|value`).
          schema:
            type: string
        - name: code
          in: query
          required: false
          description: >
            `List.code` search parameter, to filter by list/data category
            (e.g. `MedicationRequest`).
          schema:
            type: string
        - $ref: "#/components/parameters/Count"
      responses:
        "200":
          description: Search result Bundle
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIRBundle"
              example:
                resourceType: Bundle
                type: searchset
                total: 1
                entry:
                  - resource:
                      resourceType: List
                      id: "1"
                      status: current
                      mode: working
                      subject:
                        identifier:
                          system: "http://fhir.nl/fhir/NamingSystem/bsn"
                          value: "999911136"
                      source:
                        identifier:
                          system: "http://fhir.nl/fhir/NamingSystem/ura"
                          value: "00000020"
                      code:
                        coding:
                          - system: "http://minvws.github.io/generiekefuncties-docs/CodeSystem/nl-gf-data-categories-cs"
                            code: MedicationRequest
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "422":
          description: Upstream NVI response was paginated, which isn't supported
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIROperationOutcome"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
    delete:
      tags: [nvi]
      summary: Delete Lists matching search parameters
      operationId: deleteNVIListsByParams
      description: >
        Same parameter rules as the search endpoint above (at least one of
        `patient:identifier`, `subject:identifier`, `source:identifier` is
        required).
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: patient:identifier
          in: query
          required: false
          description: >
            BSN of the patient referenced by `List.subject`, as a FHIR
            token (`system|value`). Mapped to `subject:identifier` before
            forwarding to NVI.
          schema:
            type: string
          example: "http://fhir.nl/fhir/NamingSystem/bsn|999911136"
        - name: subject:identifier
          in: query
          required: false
          description: >
            BSN of the patient referenced by `List.subject`, as a FHIR
            token (`system|value`). NVI's native parameter name; equivalent
            to `patient:identifier`.
          schema:
            type: string
        - name: source:identifier
          in: query
          required: false
          description: >
            Identifier of `List.source` (the system that registered the
            List), as a FHIR token (`system|value`).
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
  /nvi/List/_search:
    post:
      tags: [nvi]
      summary: Search for Lists (form-encoded body)
      operationId: searchNVIListsForm
      description: Same behavior as `GET /nvi/List`, with parameters in the request body.
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                patient:identifier:
                  type: string
                subject:identifier:
                  type: string
                source:identifier:
                  type: string
                code:
                  type: string
                _count:
                  type: integer
                  description: >
                    FHIR `_count` search parameter: maximum number of results
                    per page. Forwarded to NVI as-is.
                  example: 20
            example:
              patient:identifier: "http://fhir.nl/fhir/NamingSystem/bsn|999911136"
      responses:
        "200":
          description: Search result Bundle
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIRBundle"
              example:
                resourceType: Bundle
                type: searchset
                total: 1
                entry:
                  - resource:
                      resourceType: List
                      id: "1"
                      status: current
                      mode: working
                      subject:
                        identifier:
                          system: "http://fhir.nl/fhir/NamingSystem/bsn"
                          value: "999911136"
                      source:
                        identifier:
                          system: "http://fhir.nl/fhir/NamingSystem/ura"
                          value: "00000020"
                      code:
                        coding:
                          - system: "http://minvws.github.io/generiekefuncties-docs/CodeSystem/nl-gf-data-categories-cs"
                            code: MedicationRequest
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "422":
          description: Upstream NVI response was paginated, which isn't supported
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIROperationOutcome"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
  /nvi/List/{id}:
    get:
      tags: [nvi]
      summary: Read a List by ID
      operationId: getNVIList
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: id
          in: path
          required: true
          description: The List resource's `id` at NVI.
          schema:
            type: string
      responses:
        "200":
          description: The List resource
          content:
            application/fhir+json:
              schema:
                $ref: "#/components/schemas/FHIRList"
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
    delete:
      tags: [nvi]
      summary: Delete a List by ID
      operationId: deleteNVIList
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: id
          in: path
          required: true
          description: The List resource's `id` at NVI.
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/OperationOutcomeError"
        "503":
          $ref: "#/components/responses/OperationOutcomeError"
        "500":
          $ref: "#/components/responses/OperationOutcomeError"
...
  /subscription/inbox/{id}/ack:
    post:
      tags: [subscription]
      summary: Acknowledge an inbox entry (Subscription Client role)
      operationId: ackSubscriptionInboxEvent
      description: >
        At-least-once semantics on the internal hop: an entry stays in the
        default inbox view until acknowledged.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Acknowledged
        "404":
          $ref: "#/components/responses/SubscriptionError"
...
components:
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: >
        The local care organization, who is performing (or on whose behalf)
        the request is being made, as a FHIR token: `<system>|<value>` where
        `<system>` must be `http://fhir.nl/fhir/NamingSystem/ura`.
      schema:
        type: string
        x-go-type: tenants.ID
        x-go-type-import:
          path: github.com/nuts-foundation/nuts-knooppunt/lib/tenants
      example: "http://fhir.nl/fhir/NamingSystem/ura|00000020"
...
  schemas:
    FHIRBundle:
      type: object
      description: >
        A FHIR R4 Bundle resource. See https://hl7.org/fhir/R4/bundle.html
        for its schema; not redefined here.
      additionalProperties: true
      x-go-type: fhir.Bundle
      x-go-type-import:
        path: github.com/zorgbijjou/golang-fhir-models/fhir-models/fhir
...